refactor(cpu): change memory callback address parameter from uint16_t to uint32_t
Expand address parameter width in memory callback functions to support
24-bit addressing for future 65816 processor implementation while
maintaining backward compatibility.

- Update function type definitions in fam65xx_types.h
- Modify MOS6502 descriptor creation functions
- Update test harnesses to mask addresses for 16-bit compatibility
- Add extended memory support in processor test runner
- Preserve existing 16-bit address space behavior with masking

+++ code/cpp/src/chip/cpu/fam65xx/fam65xx_types.h
@@ -52,12 +52,12 @@
 
 // Modern C++ function types - preferred for new code
 #include <functional>
-using fam65xx_mem_read_func = std::function<uint8_t(void* user_data, uint16_t addr, uint8_t bus_state)>;
-using fam65xx_mem_write_func = std::function<void(void* user_data, uint16_t addr, uint8_t data)>;
+using fam65xx_mem_read_func = std::function<uint8_t(void* user_data, uint32_t addr, uint8_t bus_state)>;
+using fam65xx_mem_write_func = std::function<void(void* user_data, uint32_t addr, uint8_t data)>;
 
 // Legacy C function pointer types - kept for C compatibility
-using fam65xx_mem_read_t = uint8_t (*)(void* user_data, uint16_t addr, uint8_t bus_state);
-using fam65xx_mem_write_t = void (*)(void* user_data, uint16_t addr, uint8_t data);
+using fam65xx_mem_read_t = uint8_t (*)(void* user_data, uint32_t addr, uint8_t bus_state);
+using fam65xx_mem_write_t = void (*)(void* user_data, uint32_t addr, uint8_t data);
 
 // ============================================================================
 // Enhanced Chip Descriptor (extends chip_descriptor_t for fam65xx CPUs)
